requirements.txt
PyYAML
cerberus
click